bugfix todo

+++ src/main/java/de/lecuutex/discordbot/listener/MessageReceiveListener.java
@@ -39,6 +39,8 @@ public class MessageReceiveListener extends ListenerAdapter {
                 return;
             }
 
+            if(member.getUser().isBot()) return;
+
             Message messageEmbed = new Embed("⛏ **|** TODO", message.getContentRaw(), Color.decode("#f6b93b")).send(channel);
             messageEmbed.addReaction("✅").queue();
             message.delete().queue();
